fect:    - 상품리뷰 서비스 추가           - ProductRepository.java 에 findRevuewsByProductId 추가           -프로덕트 DTO 에GetProductDetailResponse of 추가

+++ src/main/java/com/ecommerce/products/dto/response/GetLatestReview.java
@@ -9,7 +9,7 @@ import com.ecommerce.review.entity.Review;
  */
 public record GetLatestReview(
 	String userName,        // 고객명
-	Integer rating,         // 평점
+	double rating,         // 평점
 	String content,         // 리뷰 내용
 	LocalDateTime createdAt // 작성일
 ) {
